+++ client/src/App.css
@@ -1,97 +1,5 @@
-*{
-  padding: 0;
-  margin: 0;
-}
 body {
   font-family: Arial, Helvetica, sans-serif;
   background-color: #798385;
   padding: 0;
-}
-
-.header {
-  background-color: #2d293f;
-  color: rgb(27, 148, 255);
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  padding: 0px;
-  margin: 0px;
-}
-.otsikko{
-  background-color: #913549;
-  font-size: 35px;
-  text-align: left;
-  padding: 5px;
-  font-family:'Times New Roman', Times, serif
-
-}
-.headermenu {
-  text-align: right;
-  
-}
-.headermenu button{
-  width: 150px;
-  min-height: 40px;
-  border-radius: 6px;
-  margin-right: 10px;
-  color: rgb(241, 8, 86);
-  font-size: 25px;
-  font-family:'Times New Roman', Times, serif;
-  font-weight: bold;
-
-}
-.headermenu button:hover{
-  background-color: aqua;
-}
-ul {
-  list-style-type: none;
- background: #22438c;
- border-radius: 6px;
-  
-}
-ul :hover{
-  background-color: rgb(46, 46, 117);
-  color: white;
-}
-ul li {
-
-  display: inline-block;
-  position: relative;
-}
-ul li a {
-      display: block;
-      color: #b84545;
-      padding: 20px 25px;
-      text-decoration: none;
-      text-align: center;
-      font-size: 20px;
-
-}
-ul li ul.dropdownmenu {
-  background: #3c3f29;
-  display: none;
-  width: 100%;
-  position: absolute;
-  z-index: 999;
-  
-}
-
-ul li:hover ul.dropdownmenu {
-  display: block;
-}
-#category{
-  padding: 4px;
-  font-size: 15px;
-}
-#search {
-  padding: 4px;
-  font-size: 15px;
-}
-.footer {
-  background-color: #49455e;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  padding: 10px;
-  text-align: center;
 }
