+++ client/src/App.js
@@ -1,14 +1,15 @@
 import './App.css';
-import Header from './Header';
-import Footer from './Footer';
+import Header from './components/Header';
+import Footer from './components/Footer';
+import Navbar from './components/Navbar';
 function App() {
   return (
-    <div >
+    <>
       <Header />
+      <Navbar />
       <p>tänne lisää palikoita</p>
       <Footer />
-      
-    </div>
+    </>
     
   );
 };
