+++ client/src/components/Footer.js
@@ -1,3 +1,4 @@
+import './Footer.css'
 import React from "react";
 
 const Footer = () => {
