+++ client/src/components/Header.js
@@ -1,0 +1,20 @@
+import './Header.css'
+import React from "react";
+
+const Header = () => {
+
+    return (
+        <div className="header"> 
+           
+            <div className="otsikko"> MyShowMoList
+            <div className="headermenu">
+            <button type="button"> Login</button>
+            <button type="button"> Sign Up</button>
+            </div>
+            </div>
+
+            </div>
+    );
+};
+
+export default Header;
