+++ client/src/components/Navbar.css
@@ -1,0 +1,44 @@
+
+ul {
+    list-style-type: none;
+   background: #22438c;
+    
+  }
+  ul :hover{
+    background-color: rgb(46, 46, 117);
+    color: white;
+  }
+  ul li {
+  
+    display: inline-block;
+    position: relative;
+  }
+  ul li a {
+        display: block;
+        color: #b84545;
+        padding: 20px 25px;
+        text-decoration: none;
+        text-align: center;
+        font-size: 20px;
+  
+  }
+  ul li ul.dropdownmenu {
+    background: #3c3f29;
+    display: none;
+    width: 100%;
+    position: absolute;
+    z-index: 999;
+    
+  }
+  
+  ul li:hover ul.dropdownmenu {
+    display: block;
+  }
+  #category{
+    padding: 4px;
+    font-size: 15px;
+  }
+  #search {
+    padding: 4px;
+    font-size: 15px;
+  }
